Remove icon ThemeSearchPaths option on Linux
ThemeSearchPaths option was originally introduced in:
c420de0f9b ("Option to opt-out from using a Linux desktop icon theme.", 2020-01-26)

It was set on by default in:
dc6456caf9 ("Gui: Use FreeCAD supplied icons on Linux by default", 2024-06-26)

The key effect of the option is to select the FreeCAD-default icon theme.
The code also unnecessarily resets QIcon::themeSearchPaths,
":\icons" is always included in the default search paths [1],
when when using QAdwaitaDecorations causes window frame control icons
not to be displayed:

  (qt.qpa.qadwaitadecorations) Failed to find an svg icon for  "window-close-symbolic.svg"
  (qt.qpa.qadwaitadecorations) Failed to find an svg icon for  "window-minimize-symbolic.svg"
  (qt.qpa.qadwaitadecorations) Failed to find an svg icon for  "window-maximize-symbolic.svg"
  (qt.qpa.qadwaitadecorations) Failed to find an svg icon for  "window-restore-symbolic.svg"

A different default icon theme can still be chosen with the Name option.

[1] https://doc.qt.io/qt-6/qicon.html#themeSearchPaths